+++ src/ScottPlot5/ScottPlot5/Interfaces/IPanel.cs
@@ -63,4 +63,22 @@ public static class IPanelExtensions
     /// Returns true for Y axes (left and right)
     /// </summary>
     public static bool IsVertical(this IPanel panel) => !panel.IsHorizontal();
+
+    /// <summary>
+    /// Set min and max size to the same value to force this size to always be used
+    /// </summary>
+    public static void LockSize(this IPanel panel, float pixels)
+    {
+        panel.MinimumSize = pixels;
+        panel.MaximumSize = pixels;
+    }
+
+    /// <summary>
+    /// Remove all pixel size constraints
+    /// </summary>
+    public static void ResetSize(this IPanel panel, float pixels)
+    {
+        panel.MinimumSize = 0;
+        panel.MaximumSize = float.MaxValue;
+    }
 }